# openapi.yaml
openapi: 3.0.3
info:
  title: Servicio web - Periódicos v1.0
  description: >-
    Servicio web para consumir datos de periódicos. Sitio web
    [https://api.periodicos.rmgss.net](api.periodicos.rmgss.net). Servicio
    desarrollador por Waynalabs 2023 - [Código
    fuente](https://github.com/waynalabs/periodicos-ws).
  contact:
    email: rgarcia@laotra.red
  license:
    name: Licencia AGPL v3
    url: 'https://www.gnu.org/licenses/agpl-3.0.en.html'
  version: '1.0'
servers:
  - url: 'https://api.periodicos.rmgss.net/api/'
...
paths:
  /newspapers:
    get:
      tags:
        - newspapers
      summary: Listado de periódicos
      description: Listar periódicos disponibles
      operationId: newspapers
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Newspapers'
        '404':
          description: Not found
  /newspaper:
    get:
      tags:
        - newspaper
      summary: Detalles de periódico específico.
      description: Devuelve detalles generales de un periódico específico.
      operationId: newspaper
      parameters:
        - name: name
          in: query
          description: Nombre del periódico
          required: true
          explode: false
          schema:
            type: string
            example: El Deber
        - name: startDate
          in: query
          description: >-
            Para devolver detalles **sólo a partir** de esta fecha. Formato
            `AAAA/MM/DD`
          required: false
          explode: false
          schema:
            type: string
            example: '2017-01-01'
        - name: endDate
          in: query
          description: >-
            Para devolver detalles **sólo hasta** esta fecha. Formato
            `AAAA/MM/DD`
          required: false
          explode: false
          schema:
            type: string
            example: '2020-01-01'
      responses:
        '200':
          description: Detalles de periódico
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Newspaper'
        '400':
          description: Invalid status value
        '404':
          description: Not found
  /articles:
    get:
      tags:
        - articles
      summary: Datos generales de artículos de noticias.
      description: Artículos para un periódico y rango de fechas dadas.
      operationId: articles
      parameters:
        - name: name
          in: query
          description: Nombre del periódico
          required: true
          explode: false
          schema:
            type: string
            example: El Deber
        - name: startDate
          in: query
          description: >-
            Para devolver detalles **sólo a partir** de esta fecha. Formato
            `AAAA/MM/DD`
          required: false
          explode: false
          schema:
            type: string
            example: '2017-01-01'
        - name: endDate
          in: query
          description: >-
            Para devolver detalles **sólo hasta** esta fecha. Formato
            `AAAA/MM/DD`
          required: false
          explode: false
          schema:
            type: string
            example: '2020-01-01'
        - name: limit
          in: query
          description: Limita el número de resultados
          required: false
          explode: false
          schema:
            type: integer
            example: 1000
        - name: offset
          in: query
          description: 'Desplazamiento, para devolver resultados a partir de este valor.'
          required: false
          schema:
            type: integer
            example: 0
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Articles'
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
  /article:
    get:
      tags:
        - article
      summary: Retorna detalles de un artículo específico.
      description: Detalles de un artículo específico.
      operationId: article
      parameters:
        - name: id
          in: query
          description: Identificador del artículo en específico.
          required: true
          schema:
            type: integer
            example: 72803
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '404':
          description: No encontrado
  /authorSearch:
    get:
      tags:
        - author search
      summary: Búsqueda de autor por nombre.
      description: >-
        Se busca un autor en toda la base de datos para devolver posibles
        coincidencias.
      operationId: authorSearch
      parameters:
        - name: name
          in: query
          description: Nombre del autor buscado
          required: true
          schema:
            type: string
            example: juan
        - name: limit
          in: query
          description: Limita el número de resultados.
          required: false
          explode: false
          schema:
            type: integer
            example: 10
        - name: offset
          in: query
          description: 'Desplazamiento, para devolver resultados a partir de este valor.'
          required: false
          schema:
            type: integer
            example: 0
      responses:
        '200':
          description: Autores econtrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorSearchResponse'
        '405':
          description: Invalid input
  /articlesCountByDay:
    get:
      tags:
        - articles count by day
      summary: Retorna el conteo de artículos por cada día de todos los periódicos.
      description: >-
        Dada una fecha de inicio y finalización, retorna el conteo de los
        artículos publicados por periódico y día.
      operationId: articlesCountByDay
      parameters:
        - name: startDate
          in: query
          required: true
          explode: false
          schema:
            type: string
            example: '2019-01-01'
        - name: endDate
          in: query
          description: Para devolver resultados **hasta** esta fecha. Formato `AAAA/MM/DD`
          required: true
          explode: false
          schema:
            type: string
            example: '2020-01-01'
      responses:
        '200':
          description: Conteo de artículos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticlesCountByDayResponse'
        '400':
          description: Invalid input
  /topCategoriesByMonth:
    get:
      tags:
        - top categories by month
      summary: >-
        Las tres categorías con más aartículos por cada mes de in periódico
        particular.
      description: >-
        Dada una fecha de inicio y finalizacíon, retorna un conteo con las tres
        categorías con más artículos por cada mes de un periódico dado.
      operationId: topCategoriesByMonth
      parameters:
        - name: newspaper
          in: query
          description: Nombre del periódico
          required: false
          explode: false
          schema:
            type: string
            example: El Deber
        - name: startDate
          in: query
          required: true
          explode: false
          schema:
            type: string
            example: '2019-01-01'
        - name: endDate
          in: query
          description: Para devolver resultados **hasta** esta fecha. Formato `AAAA/MM/DD`
          required: true
          explode: false
          schema:
            type: string
            example: '2020-01-01'
        - name: number
          in: query
          description: Para devolver un top con este número de elementos por mes. Válido de 1 a 5.
          required: false
          explode: false
          schema:
            type: integer
            example: 3
      responses:
        '200':
          description: Conteno de categorías por mes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopCategoriesByMonthResponse'
        '400':
          description: Invalid input
  /topAuthorsByMonth:
    get:
      tags:
        - top authors by month
      summary: >-
        Los tres autores con más aartículos por cada mes de in periódico
        particular.
      description: >-
        Dada una fecha de inicio y finalizacíon, retorna un conteo con los tres
        autores con más artículos por cada mes de un periódico dado.
      operationId: topAuthorsByMonth
      parameters:
        - name: newspaper
          in: query
          description: Nombre del periódico
          required: false
          explode: false
          schema:
            type: string
            example: El Deber
        - name: startDate
          in: query
          required: true
          explode: false
          schema:
            type: string
            example: '2019-01-01'
        - name: endDate
          in: query
          description: Para devolver resultados **hasta** esta fecha. Formato `AAAA/MM/DD`
          required: true
          explode: false
          schema:
            type: string
            example: '2020-01-01'
        - name: number
          in: query
          description: Para devolver un top con este número de elementos por mes. Válido de 1 a 5.
          required: false
          explode: false
          schema:
            type: integer
            example: 3
      responses:
        '200':
          description: Conteno de autores por mes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopAuthorsByMonthResponse'
        '400':
          description: Invalid input
  /fullTextSearch:
    get:
      tags:
        - full text search
      summary: Búsqueda de texto en todos los artículos.
      description: >-
        Realiza una búsqueda rápida del texto solicitado en todos los artículos.
        Los artículos encontrados se ordenan de mayor a menor por el mejor
        puntaje de aciertos.
      operationId: fullTextSearch
      parameters:
        - name: searchText
          in: query
          description: 'Texto a buscar en todos los artículos, entre 3 a 750 caracteres.'
          required: true
          schema:
            type: string
            example: Zarate Willca
        - name: rate
          in: query
          description: >-
            Puntaje de acierto mínimo requerido, por defecto es 9. Este número
            entre 1 y 100 sirve para filtrar artículos que tengan al menos ese
            porcentaje de aciertos en la búsqueda del texto dado.
          required: false
          schema:
            type: integer
            example: 9
        - name: limit
          in: query
          description: Limita el número de resultados.
          required: false
          explode: false
          schema:
            type: integer
            example: 1000
        - name: offset
          in: query
          description: 'Desplazamiento, para devolver resultados a partir de este valor.'
          required: false
          schema:
            type: integer
            example: 0
      responses:
        '200':
          description: Entidades o nombres encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullTextSearchResponse'
        '400':
          description: Invalid input.
  /namedEntitySearch:
    get:
      tags:
        - named entity search
      summary: Búsqueda de nombre o entidad en artículos.
      description: Ayuda a buscar un nombre de persona o entidad en todos los artículos.
      operationId: namedEntitySearch
      parameters:
        - name: searchText
          in: query
          description: Texto a buscar en todos los artículos.
          required: true
          schema:
            type: string
            example: morales
        - name: limit
          in: query
          description: Limita el número de resultados.
          required: false
          explode: false
          schema:
            type: integer
            example: 1000
        - name: offset
          in: query
          description: 'Desplazamiento, para devolver resultados a partir de este valor.'
          required: false
          schema:
            type: integer
            example: 0
      responses:
        '200':
          description: Entidades o nombres encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntitySearchResponse'
        '400':
          description: Invalid input.
...
components:
  schemas:
    Newspapers:
      type: object
      properties:
        count:
          type: integer
          format: int64
          example: 3
        results:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: El Deber
              min_date:
                type: string
                example: '2015-01-01'
                description: Formato `AAAA/MM/DD`
              max_date:
                type: string
                example: '2023-04-18'
                description: Formato `AAAA/MM/DD`
    Newspaper:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: El Deber
        articles_count:
          type: integer
          format: int64
          example: 17351
        authors_count:
          type: integer
          format: int64
          example: 55
        authors:
          type: array
          items:
            type: string
            example:
              - bbc mundo
              - diego seas
              - leyla mendieta
        categories_count:
          type: integer
          example: 18973
        categories:
          type: array
          items:
            type: string
            example:
              - acero
              - minerales
              - minería
              - medio ambiente
...